# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dynasties API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ShipmentStatus =
  (typeof ShipmentStatus)[keyof typeof ShipmentStatus];

export const ShipmentStatus = {
  DRAFT: "DRAFT",
  PENDING_REVIEW: "PENDING_REVIEW",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  IN_TRANSIT: "IN_TRANSIT",
  DELIVERED: "DELIVERED",
  CLOSED: "CLOSED",
  CANCELLED: "CANCELLED",
} as const;

export interface Shipment {
  id: string;
  companyId: string;
  reference: string;
  status: ShipmentStatus;
  shipperId?: string | null;
  consigneeId?: string | null;
  portOfLoading?: string | null;
  portOfDischarge?: string | null;
  vessel?: string | null;
  voyage?: string | null;
  commodity?: string | null;
  hsCode?: string | null;
  grossWeight?: number | null;
  freightTerms?: string | null;
  bookingNumber?: string | null;
  blNumber?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ComplianceScreeningStatus =
  (typeof ComplianceScreeningStatus)[keyof typeof ComplianceScreeningStatus];

export const ComplianceScreeningStatus = {
  CLEAR: "CLEAR",
  ALERT: "ALERT",
  BLOCKED: "BLOCKED",
} as const;

export type ComplianceScreeningMatchesItem = { [key: string]: unknown };

export interface ComplianceScreening {
  id: string;
  shipmentId: string;
  status: ComplianceScreeningStatus;
  screenedParties: number;
  matchCount: number;
  matches?: ComplianceScreeningMatchesItem[];
  listsChecked?: string[];
  screenedAt?: string;
}

export type RiskScoreSubScores = { [key: string]: unknown };

export type RiskScorePrimaryRiskFactorsItem = { [key: string]: unknown };

export type RiskScoreRecommendedAction =
  (typeof RiskScoreRecommendedAction)[keyof typeof RiskScoreRecommendedAction];

export const RiskScoreRecommendedAction = {
  AUTO_APPROVE: "AUTO_APPROVE",
  OPERATOR_REVIEW: "OPERATOR_REVIEW",
  ESCALATE: "ESCALATE",
} as const;

export interface RiskScore {
  id: string;
  shipmentId: string;
  compositeScore: number;
  subScores?: RiskScoreSubScores;
  primaryRiskFactors?: RiskScorePrimaryRiskFactorsItem[];
  recommendedAction: RiskScoreRecommendedAction;
  scoredAt?: string;
}

export type InsuranceQuoteCoverageType =
  (typeof InsuranceQuoteCoverageType)[keyof typeof InsuranceQuoteCoverageType];

export const InsuranceQuoteCoverageType = {
  ALL_RISK: "ALL_RISK",
  NAMED_PERILS: "NAMED_PERILS",
  TOTAL_LOSS: "TOTAL_LOSS",
} as const;

export interface InsuranceQuote {
  id: string;
  shipmentId: string;
  coverageType: InsuranceQuoteCoverageType;
  estimatedInsuredValue: number;
  estimatedPremium: number;
  currency?: string;
  coverageRationale?: string;
  exclusions?: string[];
  confidenceScore?: number;
  quotedAt?: string;
}

export type EntityEntityType =
  (typeof EntityEntityType)[keyof typeof EntityEntityType];

export const EntityEntityType = {
  SHIPPER: "SHIPPER",
  CONSIGNEE: "CONSIGNEE",
  CARRIER: "CARRIER",
  NOTIFY_PARTY: "NOTIFY_PARTY",
  FORWARDER: "FORWARDER",
  AGENT: "AGENT",
  VENDOR: "VENDOR",
  CUSTOMER: "CUSTOMER",
} as const;

export type EntityStatus = (typeof EntityStatus)[keyof typeof EntityStatus];

export const EntityStatus = {
  VERIFIED: "VERIFIED",
  UNVERIFIED: "UNVERIFIED",
} as const;

export interface Entity {
  id: string;
  companyId: string;
  name: string;
  normalizedName?: string;
  entityType: EntityEntityType;
  status: EntityStatus;
  address?: string | null;
  city?: string | null;
  country?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type IngestedDocumentExtractionStatus =
  (typeof IngestedDocumentExtractionStatus)[keyof typeof IngestedDocumentExtractionStatus];

export const IngestedDocumentExtractionStatus = {
  PENDING: "PENDING",
  PROCESSING: "PROCESSING",
  EXTRACTED: "EXTRACTED",
  FAILED: "FAILED",
} as const;

export type IngestedDocumentExtractedData = { [key: string]: unknown } | null;

export interface IngestedDocument {
  id: string;
  companyId: string;
  emailId?: string | null;
  fileName: string;
  mimeType?: string;
  documentType: string;
  extractionStatus: IngestedDocumentExtractionStatus;
  extractedData?: IngestedDocumentExtractedData;
  createdAt: string;
}

export type IngestedEmailStatus =
  (typeof IngestedEmailStatus)[keyof typeof IngestedEmailStatus];

export const IngestedEmailStatus = {
  RECEIVED: "RECEIVED",
  PROCESSING: "PROCESSING",
  PROCESSED: "PROCESSED",
  FAILED: "FAILED",
} as const;

export interface IngestedEmail {
  id: string;
  companyId: string;
  messageId: string;
  fromAddress: string;
  toAddress: string;
  subject?: string | null;
  attachmentCount?: number;
  status: IngestedEmailStatus;
  createdAt: string;
}

export type EventMetadata = { [key: string]: unknown } | null;

export interface Event {
  id: string;
  companyId: string;
  eventType: string;
  entityType: string;
  entityId: string;
  metadata?: EventMetadata;
  createdAt: string;
}

export type ListShipments200 = {
  data: Shipment[];
};

export type GetShipment200 = {
  data: Shipment;
};

export type GetShipmentCompliance200 = {
  data: ComplianceScreening[];
};

export type GetShipmentRisk200 = {
  data?: RiskScore;
};

export type GetShipmentInsurance200 = {
  data?: InsuranceQuote;
};

export type ListEntities200 = {
  data: Entity[];
};

export type GetEntity200 = {
  data: Entity;
};

export type ListDocuments200 = {
  data: IngestedDocument[];
};

export type ListEmails200 = {
  data: IngestedEmail[];
};

export type ListEventsParams = {
  type?: string;
};

export type ListEvents200 = {
  data: Event[];
};