# lib/api-zod/src/generated/types/complianceScreening.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dynasties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceScreeningMatchesItem } from "./complianceScreeningMatchesItem";
import type { ComplianceScreeningStatus } from "./complianceScreeningStatus";

export interface ComplianceScreening {
  id: string;
  shipmentId: string;
  status: ComplianceScreeningStatus;
  screenedParties: number;
  matchCount: number;
  matches?: ComplianceScreeningMatchesItem[];
  listsChecked?: string[];
  screenedAt?: Date;
}