# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dynasties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScorePrimaryRiskFactorsItem } from "./riskScorePrimaryRiskFactorsItem";
import type { RiskScoreRecommendedAction } from "./riskScoreRecommendedAction";
import type { RiskScoreSubScores } from "./riskScoreSubScores";

export interface RiskScore {
  id: string;
  shipmentId: string;
  compositeScore: number;
  subScores?: RiskScoreSubScores;
  primaryRiskFactors?: RiskScorePrimaryRiskFactorsItem[];
  recommendedAction: RiskScoreRecommendedAction;
  scoredAt?: Date;
}