# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dynasties API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShipmentStatus } from "./shipmentStatus";

export interface Shipment {
  id: string;
  companyId: string;
  reference: string;
  status: ShipmentStatus;
  shipperId?: string | null;
  consigneeId?: string | null;
  portOfLoading?: string | null;
  portOfDischarge?: string | null;
  vessel?: string | null;
  voyage?: string | null;
  commodity?: string | null;
  hsCode?: string | null;
  grossWeight?: number | null;
  freightTerms?: string | null;
  bookingNumber?: string | null;
  blNumber?: string | null;
  createdAt: Date;
  updatedAt: Date;
}